[#1043] Made first refactor.
